Fixed the accuracy issues.
Changed the bounding box to use floats for the height and width
of the square allowing us to make the bounding box very slightly
smaller which works better with terrain collision.

+++ source/bounding_box.cpp
@@ -1,6 +1,6 @@
 #include "bounding_box.hpp"
 
-BoundingBox::BoundingBox(Vec2 position, int h, int w){
+BoundingBox::BoundingBox(Vec2 position, float h, float w){
     pos = position;
     height = h; width = w;
   }
